# testdata/contract/swagger-clojure.json
{
    "basePath": "/",
    "consumes": [
        "application/json",
        "application/x-yaml",
        "application/edn",
        "application/transit+json",
        "application/transit+msgpack"
    ],
    "definitions": {
        "AVUChangeResult": {
            "additionalProperties": false,
            "properties": {
                "path": {
                    "description": "The iRODS path of the file whose AVUs changed.",
                    "type": "string"
                },
                "user": {
                    "description": "The effective user who performed the request.",
                    "type": "string"
                }
            },
            "required": [
                "path",
                "user"
            ],
            "type": "object"
        },
        "AVUGetResult": {
            "additionalProperties": {},
            "properties": {
                "irods-avus": {
                    "description": "A list of AVUs on this object",
                    "items": {
                        "$ref": "#/definitions/AVUMap"
                    },
                    "type": "array"
                },
                "path": {
                    "description": "The iRODS path of the file whose AVUs are being listed.",
                    "type": "string"
                }
            },
            "required": [
                "irods-avus",
                "path"
            ],
            "type": "object"
        },
        "AVUMap": {
            "additionalProperties": false,
            "properties": {
                "attr": {
                    "description": "The attribute name",
                    "type": "string"
                },
                "unit": {
                    "description": "The unit associated with this value",
                    "type": "string"
                },
                "value": {
                    "description": "The value associated with this attribute",
                    "type": "string"
                }
            },
            "required": [
                "attr",
                "value",
                "unit"
            ],
            "type": "object"
        },
        "AddMetadataRequest": {
            "additionalProperties": {},
            "description": "The iRODS and Metadata AVUs to add",
            "properties": {
                "irods-avus": {
                    "description": "A list of AVUs on this object",
                    "items": {
                        "$ref": "#/definitions/AVUMap"
                    },
                    "type": "array"
                }
            },
            "type": "object"
        },
...
        "FileStat": {
            "additionalProperties": false,
            "properties": {
                "file": {
                    "$ref": "#/definitions/FileStatInfo"
                }
            },
            "required": [
                "file"
            ],
            "type": "object"
        },
        "FileStatInfo": {
            "additionalProperties": false,
            "description": "File info",
            "properties": {
                "content-type": {
                    "description": "The detected media type of the data contained in this file",
                    "type": "string"
                },
                "date-created": {
                    "description": "The date this data item was created",
                    "format": "int64",
                    "type": "integer"
                },
                "date-modified": {
                    "description": "The date this data item was last modified",
                    "format": "int64",
                    "type": "integer"
                },
                "file-size": {
                    "description": "The size in bytes of this file",
                    "format": "int64",
                    "type": "integer"
                },
                "id": {
                    "description": "The UUID of this data item",
                    "format": "uuid",
                    "type": "string"
                },
                "infoType": {
                    "description": "The type of contents in this file",
                    "type": "string"
                },
                "label": {
                    "description": "The descriptive label for this item.",
                    "type": "string"
                },
                "md5": {
                    "description": "The md5 hash of this file's contents, as calculated and saved by IRODS",
                    "type": "string"
                },
                "path": {
                    "description": "The IRODS paths to this data item",
                    "type": "string"
                },
                "permission": {
                    "description": "The requesting user's permissions on this data item",
                    "enum": [
                        "read",
                        "own",
                        "write"
                    ],
                    "type": "string"
                },
                "share-count": {
                    "description": "The number of other users this data item is shared with (only displayed to users with 'own' permissions)",
                    "format": "int64",
                    "type": "integer"
                },
                "type": {
                    "description": "The data item's type",
                    "enum": [
                        "dir",
                        "file"
                    ],
                    "type": "string"
                }
            },
            "required": [
                "infoType",
                "path",
                "date-created",
                "md5",
                "permission",
                "date-modified",
                "type",
                "file-size",
                "label",
                "id",
                "content-type"
            ],
            "type": "object"
        },
...
        "TrashPaths": {
            "additionalProperties": false,
            "properties": {
                "async-task-id": {
                    "description": "An asynchronous task ID for the delete process being handled in the background",
                    "type": "string"
                },
                "paths": {
                    "description": "A list of iRODS paths",
                    "items": {
                        "description": "A non-blank string.",
                        "type": "string"
                    },
                    "type": "array"
                },
                "trash-paths": {
                    "$ref": "#/definitions/TrashPathsMap"
                }
            },
            "required": [
                "paths",
                "trash-paths"
            ],
            "type": "object"
        },
...
        "TrashPathsMap": {
            "additionalProperties": {
                "description": "The data item's path in the trash",
                "type": "string"
            },
            "description": "A map of paths from the request to their location in the trash, if any.",
            "type": "object"
        },
...
    },
    "info": {
        "description": "Documentation for the Discovery Environment Data Info REST API",
        "title": "Discovery Environment Data Info API",
        "version": "2.8.0"
    },
    "paths": {
...
        "/admin/data/{data-id}/metadata": {
            "get": {
                "description": "List all AVUs associated with a data item. Include administrative/system iRODS AVUs.\n           This service appends the info described below in the Response Model (200) to the metadata\n           service response for the given `data-id`.\n\n#### Error Codes:\n    ERR_NOT_A_USER, ERR_NOT_READABLE\n\n#### Delegates to metadata service\n    GET /avus/{target-type}/{target-id}\nPlease see the metadata service for additional response information.",
                "parameters": [
                    {
                        "description": "The UUID assigned to the file or folder",
                        "format": "uuid",
                        "in": "path",
                        "name": "data-id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "The username of the authenticated, requesting user",
                        "in": "query",
                        "name": "user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/AVUGetResult"
                        }
                    }
                },
                "summary": "List AVUs (administrative)",
                "tags": [
                    "data-by-id"
                ]
            },
            "patch": {
                "description": "Associate iRODS and Metadata AVUs with a data item. Allow adding any AVU.\n             The info described below in the `body` parameter's Model is processed by this endpoint,\n             and what's left of the request body is forwarded to the metadata service.\n\n#### Error Codes:\n    ERR_NOT_A_USER, ERR_NOT_READABLE, ERR_DOES_NOT_EXIST, ERR_NOT_WRITEABLE, ERR_NOT_AUTHORIZED\n\n#### Delegates to metadata service\n    POST /avus/{target-type}/{target-id}\nPlease see the metadata service for additional request information.",
                "parameters": [
                    {
                        "description": "The UUID assigned to the file or folder",
                        "format": "uuid",
                        "in": "path",
                        "name": "data-id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "The username of the authenticated, requesting user",
                        "in": "query",
                        "name": "user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "The iRODS and Metadata AVUs to add",
                        "in": "body",
                        "name": "AddMetadataRequest",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AddMetadataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/AVUChangeResult"
                        }
                    }
                },
                "summary": "Add AVUs (administrative)",
                "tags": [
                    "data-by-id"
                ]
            }
        },
...
        "/data": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "description": "Uploads a file into a directory as a user, given the directory exists and is writeable but the file does not exist.\n\n#### Error Codes:\n    ERR_NOT_A_USER, ERR_EXISTS, ERR_DOES_NOT_EXIST, ERR_NOT_WRITEABLE",
                "parameters": [
                    {
                        "description": "The username of the authenticated, requesting user",
                        "in": "query",
                        "name": "user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "The destination directory for the uploaded file.",
                        "in": "query",
                        "name": "dest",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "allowEmptyValue": true,
                        "description": "",
                        "in": "formData",
                        "name": "file",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/FileStat"
                        }
                    }
                },
                "summary": "Upload a file",
                "tags": [
                    "data"
                ]
            }
        },
...
        "/data/{data-id}": {
            "delete": {
                "description": "Deletes the data item with the provided UUID.\n\n#### Error Codes:\n    ERR_NOT_A_FOLDER, ERR_DOES_NOT_EXIST, ERR_NOT_WRITEABLE, ERR_TOO_MANY_PATHS, ERR_NOT_A_USER",
                "parameters": [
                    {
                        "description": "The UUID assigned to the file or folder",
                        "format": "uuid",
                        "in": "path",
                        "name": "data-id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "The username of the authenticated, requesting user",
                        "in": "query",
                        "name": "user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/TrashPaths"
                        }
                    }
                },
                "summary": "Delete Data Item",
                "tags": [
                    "data-by-id"
                ]
            },
            "head": {
                "description": "Returns an HTTP status according to the user's access level to the data item.",
                "parameters": [
                    {
                        "description": "The UUID assigned to the file or folder",
                        "format": "uuid",
                        "in": "path",
                        "name": "data-id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "The username of the authenticated, requesting user",
                        "in": "query",
                        "name": "user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User has read permissions for given data item."
                    },
                    "403": {
                        "description": "User does not have read permissions for given data item."
                    },
                    "404": {
                        "description": "Data Item ID does not exist."
                    },
                    "422": {
                        "description": "User does not exist or an internal error occurred."
                    }
                },
                "summary": "Data Item Meta-Status",
                "tags": [
                    "data-by-id"
                ]
            },
            "put": {
                "consumes": [
                    "multipart/form-data"
                ],
                "description": "Overwrites a file as a user, given the user can write to it and the file already exists.\n\n#### Error Codes:\n    ERR_NOT_A_USER, ERR_DOES_NOT_EXIST, ERR_NOT_A_FILE, ERR_NOT_WRITEABLE",
                "parameters": [
                    {
                        "description": "The UUID assigned to the file or folder",
                        "format": "uuid",
                        "in": "path",
                        "name": "data-id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "The username of the authenticated, requesting user",
                        "in": "query",
                        "name": "user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "allowEmptyValue": true,
                        "description": "",
                        "in": "formData",
                        "name": "file",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/FileStat"
                        }
                    }
                },
                "summary": "Overwrite Contents",
                "tags": [
                    "data-by-id"
                ]
            }
        },
...
    },
    "produces": [
        "application/json",
        "application/x-yaml",
        "application/edn",
        "application/transit+json",
        "application/transit+msgpack"
    ],
    "swagger": "2.0",
...
}